DraftやMemoのもつ時間関係のプロパティをLongで持つように変更する

+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/local/db/draft/DraftEntity.kt
@@ -16,7 +16,7 @@ data class DraftEntity(
     val contents: String,
 
     @ColumnInfo(name = "start_time")
-    val startTime: Double,
+    val startTime: Long,
 
     @ColumnInfo(name = "category")
     val category: String
