+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/model/Draft.kt
@@ -4,6 +4,6 @@ data class Draft(
     val id: String,
     val title: String,
     val content: String,
-    val startTime: Double,
+    val startTime: Long,
     val category: String
 )
