+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/remote/firestore/memo/MemoEntity.kt
@@ -4,6 +4,6 @@ data class MemoEntity(
     val id: String? = null,
     val title: String? = null,
     val contents: String? = null,
-    val time: Double? = null,
+    val time: Long? = null,
     val category: String? = null
 )
